removed abstraxct split which was unecessary
